pyproject.toml
[project]
name = "review-sentiment"
version = "0.1.0"
description = "Walmart review sentiment analysis using Temporal"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.22.0",
    "aiohttp>=3.10.0",
    "boto3>=1.34.0",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["review_sentiment*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]